+++ fdb-record-layer-core/src/test/java/com/apple/foundationdb/record/provider/foundationdb/indexes/MultidimensionalIndexTest.java
@@ -1,5 +1,5 @@
 /*
- * RankIndexTest.java
+ * MultidimensionalIndexTest.java
  *
  * This source file is part of the FoundationDB open source project
  *